cli: rewire crypto interfaces
